chore: remove komenci related code in verify and identify folders
