+++ src/account/Settings.test.tsx
@@ -12,7 +12,6 @@ import { ensurePincode, navigate } from 'src/navigator/NavigationService'
 import { Screens } from 'src/navigator/Screens'
 import { removeStoredPin, setPincodeWithBiometry } from 'src/pincode/authentication'
 import { Currency } from 'src/utils/currencies'
-import { KomenciAvailable } from 'src/verify/reducer'
 import { createMockStore, flushMicrotasksQueue, getMockStackScreenProps } from 'test/utils'
 import { mockE164Number, mockE164NumberPepper } from 'test/values'
 
@@ -51,11 +50,6 @@ describe('Account', () => {
           identity: { e164NumberToSalt: { [mockE164Number]: mockE164NumberPepper } },
           stableToken: { balances: { [Currency.Dollar]: '0.00' } },
           goldToken: { balance: '0.00' },
-          verify: {
-            komenciAvailable: KomenciAvailable.Yes,
-            komenci: { errorTimestamps: [] },
-            status: {},
-          },
         })}
       >
         <Settings {...getMockStackScreenProps(Screens.Settings)} />
@@ -74,38 +68,6 @@ describe('Account', () => {
           account: {
             devModeActive: true,
             e164PhoneNumber: mockE164Number,
-          },
-          verify: {
-            komenci: { errorTimestamps: [] },
-            komenciAvailable: KomenciAvailable.Yes,
-            status: {},
-          },
-        })}
-      >
-        <Settings {...getMockStackScreenProps(Screens.Settings)} />
-      </Provider>
-    )
-    expect(tree).toMatchSnapshot()
-  })
-  it('renders correctly when verification is not possible', () => {
-    const now = Date.now()
-    let tree = render(
-      <Provider
-        store={createMockStore({
-          verify: { komenci: { errorTimestamps: [] }, status: {} },
-        })}
-      >
-        <Settings {...getMockStackScreenProps(Screens.Settings)} />
-      </Provider>
-    )
-    expect(tree).toMatchSnapshot()
-    tree = render(
-      <Provider
-        store={createMockStore({
-          verify: {
-            komenciAvailable: KomenciAvailable.Yes,
-            komenci: { errorTimestamps: [now, now, now] },
-            status: {},
           },
         })}
       >
