+++ src/app/selectors.test.ts
@@ -5,100 +5,10 @@ import {
   restoreAccountSteps,
   rewardsEnabledSelector,
   storeWipeRecoverySteps,
-  verificationPossibleSelector,
 } from 'src/app/selectors'
 import { Screens } from 'src/navigator/Screens'
 import { RootState } from 'src/redux/reducers'
-import { Currency } from 'src/utils/currencies'
-import { KomenciAvailable } from 'src/verify/reducer'
 import { getMockStoreData, RecursivePartial } from 'test/utils'
-import { mockE164Number, mockE164NumberPepper } from 'test/values'
-
-describe(verificationPossibleSelector, () => {
-  it('returns true when the number pepper is already cached', () => {
-    // cached salt
-    expect(
-      verificationPossibleSelector(
-        getMockStoreData({
-          account: { e164PhoneNumber: mockE164Number },
-          stableToken: { balances: { [Currency.Dollar]: '0' } },
-          goldToken: { balance: '0' },
-          identity: { e164NumberToSalt: { [mockE164Number]: mockE164NumberPepper } },
-          verify: { komenci: { errorTimestamps: [] }, status: {} },
-        })
-      )
-    ).toBe(true)
-  })
-  it('returns true when balance is sufficient and there are no Komenci errors', () => {
-    // balance is sufficient
-    expect(
-      verificationPossibleSelector(
-        getMockStoreData({
-          account: { e164PhoneNumber: mockE164Number },
-          stableToken: { balances: { [Currency.Dollar]: '0.01' } },
-          goldToken: { balance: '0' },
-          identity: {
-            e164NumberToSalt: {},
-          },
-          verify: { komenci: { errorTimestamps: [] }, status: {} },
-        })
-      )
-    ).toBe(true)
-    expect(
-      verificationPossibleSelector(
-        getMockStoreData({
-          account: { e164PhoneNumber: mockE164Number },
-          stableToken: { balances: { [Currency.Dollar]: '0' } },
-          goldToken: { balance: '0.005' },
-          identity: { e164NumberToSalt: {} },
-          verify: { komenci: { errorTimestamps: [] }, status: {} },
-        })
-      )
-    ).toBe(true)
-  })
-  it('returns false when balance is not sufficient and there are Komenci errors', () => {
-    const now = Date.now()
-    // balance is not sufficient
-    expect(
-      verificationPossibleSelector(
-        getMockStoreData({
-          account: { e164PhoneNumber: mockE164Number },
-          stableToken: { balances: { [Currency.Dollar]: '0.009' } },
-          goldToken: { balance: '0.004' },
-          identity: {
-            e164NumberToSalt: {},
-          },
-          verify: {
-            komenci: { errorTimestamps: [now, now, now] },
-            komenciAvailable: KomenciAvailable.Yes,
-            status: {},
-          },
-        })
-      )
-    ).toBe(false)
-  })
-
-  it('returns true when balance is not sufficient and there are < 2 Komenci errors', () => {
-    // balance is not sufficient
-    expect(
-      verificationPossibleSelector(
-        getMockStoreData({
-          account: { e164PhoneNumber: mockE164Number },
-          stableToken: { balances: { [Currency.Dollar]: '0.009' } },
-          goldToken: { balance: '0.004' },
-          identity: {
-            e164NumberToSalt: {},
-          },
-          verify: {
-            komenci: { errorTimestamps: [0, 0, 0] },
-            komenciAvailable: KomenciAvailable.Yes,
-            status: {},
-          },
-        })
-      )
-    ).toBe(true)
-  })
-})
 
 describe('registrationStepsSelector', () => {
   const registrationStepsSelectorWithMockStore = (
@@ -177,7 +87,7 @@ describe('registrationStepsSelector', () => {
       [Screens.NameAndPicture]: 1,
       [Screens.PincodeSet]: 2,
       [Screens.VerificationStartScreen]: 3,
-      [Screens.VerificationInputScreen]: 3,
+      [Screens.VerificationCodeInputScreen]: 3,
     }
 
     ;(
