+++ src/app/verificationMigration.test.ts
@@ -2,11 +2,12 @@ import { AttestationStat } from '@celo/contractkit/lib/wrappers/Attestations'
 import { expectSaga } from 'redux-saga-test-plan'
 import { call, select } from 'redux-saga-test-plan/matchers'
 import { StaticProvider } from 'redux-saga-test-plan/providers'
+import { e164NumberSelector } from 'src/account/selectors'
 import { verificationMigrationRan } from 'src/app/actions'
 import { numberVerifiedSelector, ranVerificationMigrationSelector } from 'src/app/selectors'
 import { runVerificationMigration } from 'src/app/verificationMigration'
 import { e164NumberToSaltSelector } from 'src/identity/selectors'
-import { e164NumberSelector, komenciContextSelector } from 'src/verify/reducer'
+import { komenciContextSelector } from 'src/verify/reducer'
 import { getContractKitAsync } from 'src/web3/contracts'
 import { mtwAddressSelector } from 'src/web3/selectors'
 import { mockAccount, mockE164Number, mockE164NumberPepper } from 'test/values'
