+++ src/app/verificationMigration.ts
@@ -3,12 +3,13 @@ import { AttestationStat, AttestationsWrapper } from '@celo/contractkit/lib/wrap
 import getPhoneHash from '@celo/phone-utils/lib/getPhoneHash'
 import { isAccountConsideredVerified } from '@celo/utils/lib/attestations'
 import { call, put, select } from 'redux-saga-test-plan/matchers'
+import { e164NumberSelector } from 'src/account/selectors'
 import { verificationMigrationRan } from 'src/app/actions'
 import { numberVerifiedSelector, ranVerificationMigrationSelector } from 'src/app/selectors'
 import { E164NumberToSaltType } from 'src/identity/reducer'
 import { e164NumberToSaltSelector } from 'src/identity/selectors'
 import Logger from 'src/utils/Logger'
-import { e164NumberSelector, KomenciContext, komenciContextSelector } from 'src/verify/reducer'
+import { KomenciContext, komenciContextSelector } from 'src/verify/reducer'
 import { getContractKit } from 'src/web3/contracts'
 import { mtwAddressSelector } from 'src/web3/selectors'
