+++ src/identity/saga.ts
@@ -22,7 +22,6 @@ import { AddressValidationType } from 'src/identity/reducer'
 import { revokeVerificationSaga } from 'src/identity/revoke'
 import { validateAndReturnMatch } from 'src/identity/secureSend'
 import { e164NumberToAddressSelector } from 'src/identity/selectors'
-import { reportRevealStatusSaga, startVerificationSaga } from 'src/identity/verification'
 import { recipientHasNumber } from 'src/recipients/recipient'
 import { Actions as TransactionActions } from 'src/transactions/actions'
 import Logger from 'src/utils/Logger'
@@ -91,7 +90,6 @@ export function* validateRecipientAddressSaga({
 }
 
 function* watchVerification() {
-  yield takeLatest(Actions.START_VERIFICATION, startVerificationSaga)
   yield takeLeading(Actions.REVOKE_VERIFICATION, revokeVerificationSaga)
 }
 
@@ -112,10 +110,6 @@ function* watchFetchDataEncryptionKey() {
   yield takeLeading(Actions.FETCH_DATA_ENCRYPTION_KEY, fetchDataEncryptionKeyWrapper)
 }
 
-function* watchReportRevealStatus() {
-  yield takeEvery(Actions.REPORT_REVEAL_STATUS, reportRevealStatusSaga)
-}
-
 export function* identitySaga() {
   Logger.debug(TAG, 'Initializing identity sagas')
   try {
@@ -124,7 +118,6 @@ export function* identitySaga() {
     yield spawn(watchValidateRecipientAddress)
     yield spawn(watchNewFeedTransactions)
     yield spawn(watchFetchDataEncryptionKey)
-    yield spawn(watchReportRevealStatus)
   } catch (error) {
     Logger.error(TAG, 'Error initializing identity sagas', error)
   } finally {
