--- src/identity/verification.test.ts
@@ -1,373 +1,0 @@
-import { ActionableAttestation } from '@celo/contractkit/lib/wrappers/Attestations'
-import { expectSaga } from 'redux-saga-test-plan'
-import { call, delay, select } from 'redux-saga/effects'
-import { e164NumberSelector } from 'src/account/selectors'
-import { AppEvents, VerificationEvents } from 'src/analytics/Events'
-import ValoraAnalytics from 'src/analytics/ValoraAnalytics'
-import {
-  cancelVerification,
-  reportRevealStatus,
-  resendAttestations,
-  startVerification,
-} from 'src/identity/actions'
-import { fetchPhoneHashPrivate } from 'src/identity/privateHashing'
-import {
-  AttestationCode,
-  NUM_ATTESTATIONS_REQUIRED,
-  reportActionableAttestationsStatuses,
-  reportRevealStatusSaga,
-  startVerificationSaga,
-  VERIFICATION_TIMEOUT,
-} from 'src/identity/verification'
-import {
-  fail,
-  idle,
-  KomenciAvailable,
-  setOverrideWithoutVerification,
-  shouldUseKomenciSelector,
-  start,
-  State as VerificationState,
-  succeed,
-  verificationStatusSelector,
-} from 'src/verify/reducer'
-import { getContractKitAsync } from 'src/web3/contracts'
-import { getConnectedUnlockedAccount, unlockAccount, UnlockResult } from 'src/web3/saga'
-import {
-  mockAccount,
-  mockE164Number,
-  mockE164NumberHash,
-  mockE164NumberPepper,
-  mockPublicDEK,
-} from 'test/values'
-
-const MockedAnalytics = ValoraAnalytics as any
-
-jest.mock('src/web3/saga', () => ({
-  ...(jest.requireActual('src/web3/saga') as any),
-  unlockAccount: jest.fn(),
-}))
-
-const mockUnlockAccount = unlockAccount as jest.MockedFunction<typeof unlockAccount>
-// eslint-disable-next-line require-yield
-mockUnlockAccount.mockImplementation(function* () {
-  return UnlockResult.SUCCESS
-})
-
-jest.mock('src/transactions/send', () => ({
-  sendTransaction: jest.fn(),
-  sendTransactionPromises: jest.fn(() => ({
-    confirmation: true,
-    transactionHash: true,
-    receipt: true,
-  })),
-}))
-
-jest.mock('react-native-sms-retriever', () => ({
-  startSmsRetriever: jest.fn(() => true),
-  addSmsListener: jest.fn(),
-  removeSmsListener: jest.fn(),
-}))
-
-jest.mock('@celo/utils', () => {
-  const mockParseSig = jest.fn(() => ({ r: 'r', s: 's', v: 'v' }))
-  return {
-    ...(jest.requireActual('@celo/utils') as any),
-    ECIES: { Encrypt: jest.fn(() => Buffer.from('0', 'hex')) },
-    SignatureUtils: { parseSignature: mockParseSig, parseSignatureWithoutPrefix: mockParseSig },
-  }
-})
-
-const attestationCode0: AttestationCode = {
-  code: 'ab8049b95ac02e989aae8b61fddc10fe9b3ac3c6aebcd3e68be495570b2d3da15aabc691ab88de69648f988fab653ac943f67404e532cfd1013627f56365f36501',
-  issuer: '848920b14154b6508b8d98e7ee8159aa84b579a4',
-}
-
-const attestationCode1: AttestationCode = {
-  code: '2033a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7998e1e8b183cf6b8cb4d4a8d923fecfddd191e61e074adc5e00',
-  issuer: 'fdb8da92c3597e81c2737e8be793bee9f1172045',
-}
-
-const attestationCode2: AttestationCode = {
-  code: '1930a9e1268576bf5dfee354a37480529d71f99be82c05005ffb71c7d742d10e7a9aa01f8acc4d7993f75ab183cf6b8cb4d4a8d923fecfddd191e61e074adc5a10',
-  issuer: 'ecb8da92c3597e81c2737e8be793bee9f1173156',
-}
-
-const mockContractKitTxObject = { txo: {} }
-
-const mockActionableAttestations: ActionableAttestation[] = [
-  {
-    issuer: attestationCode0.issuer,
-    blockNumber: 100,
-    attestationServiceURL: 'https://fake.celo.org/0',
-    name: '',
-    version: '1.1.0',
-  },
-  {
-    issuer: attestationCode1.issuer,
-    blockNumber: 110,
-    attestationServiceURL: 'https://fake.celo.org/1',
-    name: '',
-    version: '1.1.0',
-  },
-  {
-    issuer: attestationCode2.issuer,
-    blockNumber: 120,
-    attestationServiceURL: 'https://fake.celo.org/2',
-    name: '',
-    version: '1.1.0',
-  },
-]
-
-const mockAttestationsWrapperUnverified = {
-  getVerifiedStatus: jest.fn(() => ({
-    isVerified: false,
-    numAttestationsRemaining: mockAttestationsRemainingForUnverified,
-    total: 0,
-    completed: 0,
-  })),
-  getActionableAttestations: jest
-    .fn()
-    .mockImplementationOnce(() => [])
-    .mockImplementationOnce(() => mockActionableAttestations),
-  getUnselectedRequest: jest.fn(() => ({ blockNumber: 0 })),
-  approveAttestationFee: jest.fn(() => mockContractKitTxObject),
-  request: jest.fn(() => mockContractKitTxObject),
-  waitForSelectingIssuers: jest.fn(),
-  selectIssuers: jest.fn(() => mockContractKitTxObject),
-  findMatchingIssuer: jest.fn(() => 'mockIssuer'),
-  validateAttestationCode: jest.fn(() => true),
-  revealPhoneNumberToIssuer: jest.fn(() => ({ ok: true, statusCode: 'good', json: () => ({}) })),
-  complete: jest.fn(() => mockContractKitTxObject),
-}
-
-const mockAttestationsRemainingForUnverified = NUM_ATTESTATIONS_REQUIRED
-
-const mockAccountsWrapper = {
-  getWalletAddress: jest.fn(() => Promise.resolve(mockAccount)),
-  getDataEncryptionKey: jest.fn(() => Promise.resolve(mockPublicDEK)),
-}
-
-const mockVerificationState: VerificationState = {
-  komenci: {
-    errorTimestamps: [],
-    unverifiedMtwAddress: null,
-    sessionActive: false,
-    sessionToken: '',
-    callbackUrl: undefined,
-    captchaToken: '',
-  },
-  status: {
-    isVerified: false,
-    numAttestationsRemaining: 3,
-    total: 0,
-    completed: 0,
-    komenci: true,
-  },
-  actionableAttestations: [],
-  retries: 0,
-  currentState: idle(),
-  komenciAvailable: KomenciAvailable.Unknown,
-  TEMPORARY_override_withoutVerification: undefined,
-  phoneHash: mockE164NumberHash,
-  withoutRevealing: false,
-  e164Number: mockE164Number,
-  komenciConfig: {
-    useLightProxy: false,
-    allowedDeployers: [],
-  },
-}
-
-beforeEach(() => {
-  jest.clearAllMocks()
-})
-
-describe(startVerificationSaga, () => {
-  beforeEach(() => {
-    MockedAnalytics.track.mockReset()
-  })
-  it('succeeds', async () => {
-    const withoutRevealing = false
-    await expectSaga(startVerificationSaga, { withoutRevealing })
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [select(shouldUseKomenciSelector), true],
-
-        [call(reportActionableAttestationsStatuses), null],
-      ])
-      .put(setOverrideWithoutVerification(withoutRevealing))
-      .put(start({ e164Number: mockE164Number, withoutRevealing }))
-      .dispatch(succeed())
-      .run()
-    expect(MockedAnalytics.track.mock.calls.length).toBe(2)
-    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(VerificationEvents.verification_start)
-    expect(MockedAnalytics.track.mock.calls[1][0]).toBe(VerificationEvents.verification_complete)
-  })
-  it('restarts', async () => {
-    const withoutRevealing = true
-    const numAttestationsRemaining = 3
-    await expectSaga(startVerificationSaga, { withoutRevealing })
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [select(verificationStatusSelector), { numAttestationsRemaining }],
-        [select(shouldUseKomenciSelector), true],
-        [call(reportActionableAttestationsStatuses), null],
-      ])
-      .put(setOverrideWithoutVerification(withoutRevealing))
-      .put(start({ e164Number: mockE164Number, withoutRevealing }))
-      .put(startVerification(mockE164Number, false))
-      .dispatch(resendAttestations())
-      .run()
-    expect(MockedAnalytics.track.mock.calls.length).toBe(2)
-    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(VerificationEvents.verification_start)
-    expect(MockedAnalytics.track.mock.calls[1][0]).toBe(
-      VerificationEvents.verification_resend_messages
-    )
-    expect(MockedAnalytics.track.mock.calls[1][1]).toStrictEqual({
-      count: numAttestationsRemaining,
-      feeless: true,
-    })
-  })
-  it('tracks failure', async () => {
-    const withoutRevealing = false
-    const errorMessage = 'This is an error message'
-    await expectSaga(startVerificationSaga, { withoutRevealing })
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [select(shouldUseKomenciSelector), true],
-        [call(reportActionableAttestationsStatuses), null],
-      ])
-      .put(setOverrideWithoutVerification(withoutRevealing))
-      .put(start({ e164Number: mockE164Number, withoutRevealing }))
-      .dispatch(fail(errorMessage))
-      .run()
-    expect(MockedAnalytics.track.mock.calls.length).toBe(2)
-    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(VerificationEvents.verification_start)
-    expect(MockedAnalytics.track.mock.calls[1][0]).toBe(VerificationEvents.verification_error)
-    expect(MockedAnalytics.track.mock.calls[1][1]).toStrictEqual({
-      error: errorMessage,
-      feeless: true,
-    })
-  })
-
-  it('times out when verification takes too long', async () => {
-    await expectSaga(startVerificationSaga, { withoutRevealing: false })
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [select(shouldUseKomenciSelector), true],
-        [delay(VERIFICATION_TIMEOUT), true],
-        [call(reportActionableAttestationsStatuses), null],
-      ])
-      .run(2000)
-    expect(MockedAnalytics.track.mock.calls.length).toBe(3)
-    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(VerificationEvents.verification_start)
-    expect(MockedAnalytics.track.mock.calls[1][0]).toBe(VerificationEvents.verification_timeout)
-    expect(MockedAnalytics.track.mock.calls[2][0]).toBe(AppEvents.error_displayed)
-  })
-
-  it('stops when the user cancels', async () => {
-    await expectSaga(startVerificationSaga, { withoutRevealing: false })
-      .provide([
-        [select(e164NumberSelector), mockE164Number],
-        [select(shouldUseKomenciSelector), true],
-        [call(reportActionableAttestationsStatuses), null],
-      ])
-      .dispatch(cancelVerification())
-      .run(2000)
-    expect(MockedAnalytics.track.mock.calls.length).toBe(2)
-    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(VerificationEvents.verification_start)
-    expect(MockedAnalytics.track.mock.calls[1][0]).toBe(VerificationEvents.verification_cancel)
-  })
-})
-
-describe(reportActionableAttestationsStatuses, () => {
-  it('report actionable attestations', async () => {
-    const contractKit = await getContractKitAsync()
-    const mockAttestationsWrapper = {
-      ...mockAttestationsWrapperUnverified,
-      getActionableAttestations: jest.fn(() => mockActionableAttestations),
-    }
-    await expectSaga(reportActionableAttestationsStatuses)
-      .provide([
-        [call(getConnectedUnlockedAccount), mockAccount],
-        [select(e164NumberSelector), mockE164Number],
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-        [call([contractKit.contracts, contractKit.contracts.getAccounts]), mockAccountsWrapper],
-        [
-          call(fetchPhoneHashPrivate, mockE164Number),
-          {
-            phoneHash: mockE164NumberHash,
-            e164Number: mockE164Number,
-            pepper: mockE164NumberPepper,
-          },
-        ],
-      ])
-      .put(
-        reportRevealStatus(
-          mockActionableAttestations[0].attestationServiceURL,
-          mockAccount,
-          mockActionableAttestations[0].issuer,
-          mockVerificationState.e164Number!,
-          mockE164NumberPepper
-        )
-      )
-      .put(
-        reportRevealStatus(
-          mockActionableAttestations[1].attestationServiceURL,
-          mockAccount,
-          mockActionableAttestations[1].issuer,
-          mockVerificationState.e164Number!,
-          mockE164NumberPepper
-        )
-      )
-      .put(
-        reportRevealStatus(
-          mockActionableAttestations[2].attestationServiceURL,
-          mockAccount,
-          mockActionableAttestations[2].issuer,
-          mockVerificationState.e164Number!,
-          mockE164NumberPepper
-        )
-      )
-      .run()
-  })
-})
-
-describe(reportRevealStatusSaga, () => {
-  beforeEach(() => {
-    MockedAnalytics.track.mockReset()
-  })
-  it('report actionable attestation to analytics', async () => {
-    const contractKit = await getContractKitAsync()
-    const mockAttestationsWrapper = {
-      ...mockAttestationsWrapperUnverified,
-      getRevealStatus: jest.fn(() => ({
-        ok: true,
-        json: () => body,
-      })),
-    }
-    const mockIssuer = mockActionableAttestations[0].issuer
-    const body = { issuer: mockIssuer, custom: 'payload' }
-    await expectSaga(reportRevealStatusSaga, {
-      attestationServiceUrl: 'url',
-      e164Number: mockE164Number,
-      account: mockAccount,
-      issuer: mockIssuer,
-      pepper: mockE164NumberPepper,
-    })
-      .provide([
-        [
-          call([contractKit.contracts, contractKit.contracts.getAttestations]),
-          mockAttestationsWrapper,
-        ],
-      ])
-      .run()
-    expect(MockedAnalytics.track.mock.calls.length).toBe(1)
-    expect(MockedAnalytics.track.mock.calls[0][0]).toBe(
-      VerificationEvents.verification_reveal_attestation_status
-    )
-    expect(MockedAnalytics.track.mock.calls[0][1]).toStrictEqual(body)
-  })
-})
